requirements.txt
crewai==1.1.0
crewai-tools==1.1.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.1.1
pandas>=2.2.0
openpyxl>=3.1.5
requests>=2.32.5
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2
httpx>=0.27.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
gspread==5.12.4
oauth2client==4.1.3
imapclient==2.3.1
email-validator==2.1.0
cryptography>=41.0.8
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
sqlalchemy>=2.0.23
alembic>=1.13.1
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.4
streamlit==1.28.1
plotly==5.17.0
anthropic>=0.18.0
